Made astar.cpp heuristic function available to python.
Removed astar.py since I'm sticking with the C++ implementation.

Rearranged test suite a bit so conftest.py is in the root dir. May change it again later on.
