+++ hades/extensions/__init__.py
@@ -1,5 +1,8 @@
 """Contains all the extensions used by the game to speed up various features."""
 
-from hades.extensions.astar.astar import calculate_astar_path
+from hades.extensions.astar.astar import calculate_astar_path, heuristic
 
-__all__ = ("calculate_astar_path",)
+__all__ = (
+    "calculate_astar_path",
+    "heuristic",
+)
