+++ hades/extensions/astar/astar.pyi
@@ -11,3 +11,4 @@ from hades.generation.primitives import Point
 def calculate_astar_path(
     grid: np.ndarray, start: Point, end: Point, obstacle_id: TileType
 ) -> list[tuple[int, int]]: ...
+def heuristic(a: Point, b: Point) -> int: ...
