+++ tests/extensions/test_setup.py
@@ -1,0 +1,12 @@
+"""Tests all functions in extensions/setup.py."""
+from __future__ import annotations
+
+# Pip
+import pytest
+
+__all__ = ()
+
+
+def test_setup_script() -> None:
+    """Test the setup.py script."""
+    pytest.fail("IMPROVE SETUP.PY")
